first go

+++ src/application/dashboard/service.ts
@@ -1,27 +1,41 @@
 import { SubmissionId } from '../../domain/submission/types';
 import Submission from '../../domain/submission/services/models/submission';
 import { SubmissionService } from 'src/domain/submission';
+import { PermissionService, SubmissionOperation } from '../permission/service';
+import { User } from 'src/domain/user/user';
 
 export class DashboardService {
-    constructor(private readonly submissionService: SubmissionService) {}
+    constructor(
+        private readonly submissionService: SubmissionService,
+        private readonly permissionService: PermissionService,
+    ) {}
 
-    async find(): Promise<Submission[]> {
-        // do some security and validation here
-        return this.submissionService.findAll();
+    async findMySubmissions(user: User): Promise<Submission[]> {
+        return this.submissionService.findByUserId(user.id);
     }
 
-    async startSubmission(articleType: string, userId: string): Promise<Submission> {
-        // do some security and validation here
-        return this.submissionService.create(articleType, userId);
+    async startSubmission(user: User, articleType: string): Promise<Submission> {
+        const allowed = this.permissionService.userCan(user, SubmissionOperation.CREATE, null);
+        if (!allowed) {
+            throw new Error('User not allowed to create submission');
+        }
+        return this.submissionService.create(articleType, user.id);
     }
 
-    async getSubmission(id: SubmissionId): Promise<Submission> {
-        // do some security and validation here
+    async getSubmission(user: User, id: SubmissionId): Promise<Submission> {
+        const allowed = this.permissionService.userCan(user, SubmissionOperation.READ, null);
+        if (!allowed) {
+            throw new Error('User not allowed to read submission');
+        }
         return this.submissionService.get(id);
     }
 
-    async deleteSubmission(id: SubmissionId): Promise<boolean> {
-        // do some security and validation here
+    async deleteSubmission(user: User, id: SubmissionId): Promise<boolean> {
+        const allowed = this.permissionService.userCan(user, SubmissionOperation.DELETE, null);
+        if (!allowed) {
+            throw new Error('User not allowed to delete submission');
+        }
+
         return this.submissionService.delete(id);
     }
 }
